move Logger detail function defintions to src file
